+++ js/main.js
@@ -4,7 +4,7 @@ const MAX_CARDS = 10
 import './validation.js'
 import {setFormActivity} from'./form.js'
 import {setFilter, onFilterChanges} from './filter.js'
-import {getData} from './backend.js'
+import {getData} from './api.js'
 import {appendCardsToMap, initEmptyMap} from './map.js'
 import {showMessage} from './utils.js'
